requirements.txt
langchain
python-dotenv==1.0.1
requests
flask
flask-sqlalchemy
flask-migrate
gitpython
psutil
redis==5.0.4
flask-socketio==5.3.6
psycopg2-binary==2.9.9
pgvector
sentence-transformers
tiktoken
numpy
prometheus_client
PyJWT
prompt-toolkit==3.0.43
pyyaml==6.0.1
python-engineio==4.9.1
python-socketio==5.11.2
gevent>=23.9.1
gevent-websocket>=0.10.1
